# openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/compliance_definitions": {
            "post": {
                "operationId": "ComplianceController_create",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Compliance Definitions"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid"
                        ]
                    }
                ]
            },
            "get": {
                "operationId": "ComplianceController_findAll",
                "parameters": [
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all compliance definitions"
                    }
                },
                "tags": [
                    "Compliance Definitions"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid"
                        ]
                    }
                ]
            }
        },
        "/compliance_definitions/{name}": {
            "get": {
                "operationId": "GetComplianceDefinitionByName",
                "summary": "Get by name",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "path",
                        "example": "{{test_compliance_definition_name}}",
                        "description": "The name of compliance definition to return",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all compliance definitions by type"
                    }
                },
                "tags": [
                    "Compliance Definitions"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid"
                        ]
                    }
                ]
            },
            "patch": {
                "operationId": "UpdateComplianceDefinitionByName",
                "summary": "Update by name",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "path",
                        "example": "{{test_compliance_definition_name}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Compliance Definitions"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid"
                        ]
                    }
                ]
            },
            "delete": {
                "operationId": "ComplianceController_remove",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "path",
                        "example": "{{test_compliance_definition_name}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                },
                "tags": [
                    "Compliance Definitions"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid"
                        ]
                    }
                ]
            }
        },
        "/compliance_definitions/{name}/organization/{orgId}": {
            "post": {
                "operationId": "ComplianceController_createOrgCompliance",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "required": true,
                        "in": "path",
                        "example": "{{test_compliance_definition_name}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Compliance Definitions"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid"
                        ]
                    }
                ]
            },
            "delete": {
                "operationId": "ComplianceController_deactivateCompliance",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "required": true,
                        "in": "path",
                        "example": "{{test_compliance_definition_name}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Compliance Definitions"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid"
                        ]
                    }
                ]
            }
        },
        "/compliance_definitions/organization/{orgId}": {
            "get": {
                "operationId": "ComplianceController_getOrgComp",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all activated compliances for an organization"
                    }
                },
                "tags": [
                    "Compliance Definitions"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid"
                        ]
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "V1 REST API",
        "description": "🌱 This is a restful api for the cold platform 🌱 ",
        "version": "0.0.0",
        "contact": {},
        "termsOfService": "http://www.coldclimate.com/tos"
    },
...
    "servers": [
        {
            "url": "https://api.coldclimate.com/v1",
            "description": "Production Server"
        },
        {
            "url": "https://api.coldclimate.online/v1",
            "description": "Staging Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "https://dev-6qt527e13qyo4ls6.us.auth0.com/authorize?audience=https://api.coldclimate.wtf/",
                        "scopes": {
                            "all": "openid profile email"
                        }
                    }
                }
            }
        },
...
    }
}